requirements.txt
# Install a matching PyTorch / torchvision build for your CUDA version first.
accelerate==0.30.1
deepspeed==0.14.2
diffusers==0.27.2
huggingface_hub==0.25.2
imgaug==0.4.0
matplotlib>=3.8,<3.11
numpy>=1.26,<2
packaging==24.0
pillow==10.2.0
pyyaml==6.0.1
sentencepiece==0.2.0
timm==1.0.3
transformers==4.41.0
wandb==0.17.0
